Rename thing_server module to server
Otherwise,`thing_` is going to appear everywhere and
become a bit redundant...
