+++ examples/opencv_camera_server.py
@@ -6,7 +6,7 @@ from fastapi.responses import HTMLResponse, StreamingResponse
 from labthings_fastapi.descriptors.property import PropertyDescriptor
 from labthings_fastapi.thing import Thing
 from labthings_fastapi.decorators import thing_action, thing_property
-from labthings_fastapi.thing_server import ThingServer
+from labthings_fastapi.server import ThingServer
 from labthings_fastapi.file_manager import FileManagerDep
 from typing import Optional, AsyncContextManager
 from collections.abc import AsyncGenerator
