+++ tests/test_action_logging.py
@@ -4,7 +4,7 @@ This tests the log that is returned in an action invocation
 
 import logging
 from fastapi.testclient import TestClient
-from labthings_fastapi.thing_server import ThingServer
+from labthings_fastapi.server import ThingServer
 from temp_client import poll_task
 from labthings_fastapi.thing import Thing
 from labthings_fastapi.decorators import thing_action
