+++ tests/test_action_manager.py
@@ -1,7 +1,7 @@
 from fastapi.testclient import TestClient
 import pytest
 import httpx
-from labthings_fastapi.thing_server import ThingServer
+from labthings_fastapi.server import ThingServer
 from temp_client import poll_task
 import time
 from labthings_fastapi.thing import Thing
