+++ tests/test_properties.py
@@ -2,7 +2,7 @@ from labthings_fastapi.descriptors import PropertyDescriptor
 from labthings_fastapi.decorators import thing_property, thing_action
 from labthings_fastapi.thing import Thing
 from fastapi.testclient import TestClient
-from labthings_fastapi.thing_server import ThingServer
+from labthings_fastapi.server import ThingServer
 from threading import Thread
 from pytest import raises
 from pydantic import BaseModel
