+++ tests/test_thing_dependencies.py
@@ -6,7 +6,7 @@ import inspect
 from fastapi.testclient import TestClient
 from fastapi import Request
 import pytest
-from labthings_fastapi.thing_server import ThingServer
+from labthings_fastapi.server import ThingServer
 from temp_client import poll_task
 from labthings_fastapi.thing import Thing
 from labthings_fastapi.decorators import thing_action
