+++ tests/test_websocket.py
@@ -1,5 +1,5 @@
 from fastapi.testclient import TestClient
-from labthings_fastapi.thing_server import ThingServer
+from labthings_fastapi.server import ThingServer
 from test_thing import MyThing
 
 my_thing = MyThing()
